requirements.txt
fastapi
uvicorn
python-dotenv
twilio
openai
python-multipart